+++ test/deprecated/legacy_test/CMakeLists.txt
@@ -454,7 +454,7 @@ endif()
 # Some ops need to check results when gc is enabled
 # Currently, only ops that register NoNeedBufferVarsInference need to do this test
 set(TEST_OPS_WITH_GC test_affine_channel_op test_gather_nd_op test_lod_reset_op
-                     test_lookup_table_op test_scatter_op test_slice_op)
+                     test_scatter_op test_slice_op)
 
 foreach(TEST_OP ${TEST_OPS_WITH_GC})
   list(REMOVE_ITEM TEST_OPS ${TEST_OP})
@@ -825,8 +825,6 @@ set(STATIC_BUILD_TESTS
     test_fetch_lod_tensor_array
     test_fuse_bn_act_pass
     test_layer_norm_op
-    test_lookup_table_bf16_op
-    test_lookup_table_v2_op
     test_matmul_op
     test_matmul_v2_op
     test_momentum_op
