+++ test/legacy_test/test_lookup_table_op.py
@@ -19,13 +19,11 @@ from op import Operator
 from op_test import (
     OpTest,
     check_out_dtype,
-    paddle_static_guard,
     skip_check_grad_ci,
 )
 
-import paddle
 import paddle.nn.functional as F
-from paddle.base import Program, core, program_guard
+from paddle.base import core
 
 
 class TestLookupTableOp(OpTest):
@@ -157,46 +155,6 @@ class TestLookupTableWithTensorIdsWIsSelectedRows(
     def check_result(self, ids_array, result_array):
         for idx, row in np.ndenumerate(ids_array):
             assert (row == result_array[idx]).all()
-
-
-class TestEmbedOpError(unittest.TestCase):
-    def test_errors(self):
-        with paddle_static_guard():
-            with program_guard(Program(), Program()):
-                input_data = np.random.randint(0, 10, (4, 1)).astype("int64")
-
-                def test_Variable():
-                    # the input type must be Variable
-                    paddle.static.nn.embedding(input=input_data, size=(10, 64))
-
-                self.assertRaises(TypeError, test_Variable)
-
-                def test_input_dtype():
-                    # the input dtype must be int64
-                    input = paddle.static.data(
-                        name='x', shape=[4, 1], dtype='float32'
-                    )
-                    paddle.static.nn.embedding(input=input, size=(10, 64))
-
-                self.assertRaises(TypeError, test_input_dtype)
-
-                def test_param_dtype():
-                    # dtype must be float32 or float64
-                    input2 = paddle.static.data(
-                        name='x2', shape=[4, 1], dtype='int64'
-                    )
-                    paddle.static.nn.embedding(
-                        input=input2, size=(10, 64), dtype='int64'
-                    )
-
-                self.assertRaises(TypeError, test_param_dtype)
-
-                input3 = paddle.static.data(
-                    name='x3', shape=[4, 1], dtype='int64'
-                )
-                paddle.static.nn.embedding(
-                    input=input3, size=(10, 64), dtype='float16'
-                )
 
 
 class TestLookupTableOpInt8(OpTest):
